// web/src/api/generated/models/DateRepeat.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * How often an important date recurs.
 * 
 * Deliberately limited to two values. A full recurrence rule belongs to
 * later reminder logic; birthdays and anniversaries do not need it.
 * @export
 */
export const DateRepeat = {
    NONE: 'NONE',
    ANNUALLY: 'ANNUALLY'
} as const;
export type DateRepeat = typeof DateRepeat[keyof typeof DateRepeat];


export function instanceOfDateRepeat(value: any): boolean {
    for (const key in DateRepeat) {
        if (Object.prototype.hasOwnProperty.call(DateRepeat, key)) {
            if (DateRepeat[key as keyof typeof DateRepeat] === value) {
                return true;
            }
        }
    }
    return false;
}

export function DateRepeatFromJSON(json: any): DateRepeat {
    return DateRepeatFromJSONTyped(json, false);
}

export function DateRepeatFromJSONTyped(json: any, ignoreDiscriminator: boolean): DateRepeat {
    return json as DateRepeat;
}

export function DateRepeatToJSON(value?: DateRepeat | null): any {
    return value as any;
}

export function DateRepeatToJSONTyped(value: any, ignoreDiscriminator: boolean): DateRepeat {
    return value as DateRepeat;
}