// web/src/api/generated/models/ImportantDateType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ImportantDateType = {
    BIRTHDAY: 'BIRTHDAY',
    ANNIVERSARY: 'ANNIVERSARY',
    CUSTOM: 'CUSTOM'
} as const;
export type ImportantDateType = typeof ImportantDateType[keyof typeof ImportantDateType];


export function instanceOfImportantDateType(value: any): boolean {
    for (const key in ImportantDateType) {
        if (Object.prototype.hasOwnProperty.call(ImportantDateType, key)) {
            if (ImportantDateType[key as keyof typeof ImportantDateType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ImportantDateTypeFromJSON(json: any): ImportantDateType {
    return ImportantDateTypeFromJSONTyped(json, false);
}

export function ImportantDateTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImportantDateType {
    return json as ImportantDateType;
}

export function ImportantDateTypeToJSON(value?: ImportantDateType | null): any {
    return value as any;
}

export function ImportantDateTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ImportantDateType {
    return value as ImportantDateType;
}