// web/src/api/generated/models/RelatedPersonView.ts
export interface RelatedPersonView {
    /**
     * 
     * @type {string}
     * @memberof RelatedPersonView
     */
    avatarAttachmentId?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof RelatedPersonView
     */
    birthday: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof RelatedPersonView
     */
    birthdayYearKnown: boolean;
    /**
     * 
     * @type {Date}
     * @memberof RelatedPersonView
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof RelatedPersonView
     */
    displayName: string;
    /**
     * 
     * @type {string}
     * @memberof RelatedPersonView
     */
    id: string;
    /**
     * 
     * @type {PersonRelationship}
     * @memberof RelatedPersonView
     */
    relationship: PersonRelationship;
    /**
     * 
     * @type {boolean}
     * @memberof RelatedPersonView
     */
    showBirthdayOnDashboard: boolean;
    /**
     * 
     * @type {Date}
     * @memberof RelatedPersonView
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof RelatedPersonView
     */
    version: number;
    /**
     * 
     * @type {ContentVisibility}
     * @memberof RelatedPersonView
     */
    visibility: ContentVisibility;
}